fix(logging): cast the size arguments the macOS toolchain types differently
The format attributes added in #216, and the `z` coverage added in #332, catch
mismatches that only one of the two cross toolchains reports. `sizeof` and `%z`
are checked against the compiler's `__SIZE_TYPE__`, which is `unsigned int` for
the i686-elf gcc the Ubuntu job uses and `long unsigned int` for the one
Homebrew ships. Both are 32 bits here, so only the type name differs and only
the format check notices. The macOS job has been failing on `develop` since
those landed:

    heap.c:342:17: error: format '%u' expects argument of type 'unsigned int',
    but argument 7 has type 'long unsigned int' [-Werror=format=]

    t_printf_length.c:72:28: error: format '%zx' expects argument of type
    'size_t', but argument 4 has type 'unsigned int' [-Werror=format=]

The `sizeof` arguments are cast to `unsigned`, which matches `%u` on either
toolchain. The `%z` cases cannot be fixed that way, because `%z` is the one
conversion whose expected type is exactly the one the toolchains disagree on:
they use `__SIZE_TYPE__` and `__PTRDIFF_TYPE__`, so each toolchain checks them
against its own spelling. `stddef.h` declares `size_t` as `unsigned long`,
which matches the Homebrew toolchain and not the other one; leaving that alone
here, since changing it would move every format check that involves a `size_t`.

The sites were found two ways, since neither alone is complete: preprocessing
every translation unit of the local build with its own flags, which catches a
`sizeof` that arrives through a macro, and scanning the sources, which catches
a translation unit the local build does not configure.

- `heap.c:342`, where the `sizeof` arrives through the `OVERHEAD` macro.
- `t_printf_length.c`, in the `z` cases of the length-modifier coverage.
- `t_big_write.c:39`, which is built but not in the list `runtests` runs.
- `multiboot.c:162` and `ext2_dir.c`, in `pr_debug` calls that the debug level
  of those files compiles out, so they are latent rather than breaking:
  raising it would break the macOS build too. `multiboot.c` also used `%d` for
  an unsigned count, which is now `%u`.

Four more `%z` sites, in `test_utils.h`, `test_gdt.c`, `test_idt.c` and a
`pr_debug` in `zone_allocator.c`, are not in this build and are left alone.

+++ kernel/src/fs/ext2/ext2_dir.c
@@ -331,7 +331,7 @@ static inline int ext2_append_new_direntry(
         pr_debug(
             "Found last directory entry (offset: %u, %u != "
             "round(%u+%u+1):\n",
-            it.block_offset, it.direntry->rec_len, sizeof(ext2_dirent_t), it.direntry->name_len);
+            it.block_offset, it.direntry->rec_len, (unsigned)sizeof(ext2_dirent_t), it.direntry->name_len);
 
         ext2_dump_dirent(it.direntry);
         // Fix the rec_len of the entry.
